Issue #40: Fixed test case issues
To check the performance I'm going to use server side testing; so I'd to
fix those since some of django's own tests were producing to much nose.

Plus I hope that the fixtures are finally fixed: There the problem was
that my notex/fixture was containing content_type data, which is
apparently to be ignored.