Issue #40: Only every 60s new PDFs
The idea is to restrict PDF generation to 60s (or 30s, 15s?) such that
if webserver/proxy cache fails to set expiry date correctly namely
withing 60s (or lower) then the application cache at least save the day.

But setting or deleting keys with an expiry does not work with
*memcached* ;( Therefore I think it is justifiable to replace it with
*reddis*, as part of this issue.

Further since memcached deals only with object_uri's (path of the PDF)
we have to get rid of *os.remove* (shouldn't it be anyway *os.unlink*?)
and introduce a cronjob to regularly (every 60s) cleanup media.notex.ch.

Let's see if this is a viable solution.