Issue #107 Avoid *debug* tag trouble in templates
If I use lighttpd then the `debug` tag seems to be a boolean with a proper
value if system is in DEBUG mode. But if I use cherokee then the `debug`
template tag seem not to be set.

Tweaking around with INTERNAL_IPS but did not help; the strange thing is
that this *used* to work with cherokee before; very strange.

Solved using work around with a custom `dbg` tag, which is much more
explicit and pythonic in my opinion anyway (since set by *me* and not some hidden
template context processor by simply assigning the value of
setting.DEBUG).